requirements.txt
mem0ai>=0.0.1
openai>=1.0.0
qdrant-client>=1.7.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0
